ADM-2.6 r7745 bitbuffer.h.diff
Patch fixes last Avidemux-2.6 compiler error with Clang on OSX Lion.
bitbuffer.h and bitbuffer.c define the same variable with the
keyword 'inline', but to llvm that's a suggestion and to clang
that is an error, non-strict C99, as described in the llvm webpage,
'Language Compatibility', sec. 'C99 inline fucntions', here:

  http://clang.llvm.org/compatibility.html

One preferred fix is used here, removing the keyword 'inline'
from one definition, leaving the other intact.  This is p1 patch.