adm2.5 - ADM_mangle.h mark i386 array as ((used))
avidemux/ADM_core/include/ADM_mangle.h has a section for x64
and a section for i386.  The x64 section was previously patched
to stop symbols from being optimized out by the compiler.

This patch adds the same syntax for the section that handles
__APPLE__ && !ADM_CPU_X86_64

The error report is here:
http://www.avidemux.org/smf/index.php?topic=10441.0

Previous patch for x64 was this:
http://svn.berlios.de/viewvc/avidemux?view=revision&revision=7712